Fixes unicode logging decoding issue
Logging fails with an exception when unicode conversion of the message
cannot be performed.

This includes also Exception objects, where the exception message has
an unknown encoding.

The calling code should be responsible for handling the unicode encoding, but
nevertheless, the logging module should handle this scenario and not raise
an exception.

Closes-Bug: #1393870

Change-Id: Ied46e4a8715d11a3cf59d651004e17f6fe4c760b