Add testutils.ConfPatcher and use it in tests
The tests were modifying a global ConfigOpts object, without
resetting the old value, which could affect other tests as well.
The new ConfPatcher class can be used both as a context manager
and as a decorator, making sure that the old value of configuration
option is set back after exiting from the decorated function or from
the context manager.

Change-Id: If23bf225207977e0e313dc806d53bca8a40215d6