Enable password set in case of rescued instances
In the case of rescued instances, the user password has already
been set for an instance, and it cannot be posted anymore to the
http metadata service. In this case we still have to set the new
password and warn the user that the password has already been
posted to the metadata service and it cannot be changed anymore.

Change-Id: I9b52af6bf8bac2c394bc4a9eb30a983594fa1e68
Closes-Bug: #1391818