Handle the choice of reboot of the instance in osutils.set_host_name
Since the sethostname plugin was moved in plugins.common, there was a bit
of Windows specific functionality inside it, the fact that it returns reboot_required
in order to set the hostname on Windows. Now this choice is moved to
osutils.set_host_name, which will simply return False for non-Windows systems.

Change-Id: Ic7774f2c7dd86f3ade82a19520df4301bc9674a7