Add support for VFAT ConfigDrive
In order to provide support for VFAT ConfigDrive, this patch uses a couple
of tools from the mtools package, mdir for listing VFAT filesystems, respectively
mcopy, for copying files from such filesystems. The first step when retrieving the
metadata from a ConfigDrive is now the lookup for VFAT filesystems. If any such filesystem
is found, the metadata will be taken from there.

Change-Id: I2fa9c96ce22936ca3d8d8fb9f55ca1ab29ce1d1d