Add temporarily tzlocal to requirements
It seems the change id 30c56082bb9b9bcd80128eedcef17b6beb793cf8
wasn't released in pbr, so this patch adds the tzlocal module
requirement, temporarily, to requirements.txt, until a due release
of pbr will be done. There's no need to add the rest of the libraries,
because they are already provided by the installer.

Change-Id: I03c66bd13ad04e9be2164d8e450cd7c625c731dd