Get credentials from metadata for WinRM auth plugin
ConfigWinRMCertificateAuthPlugin plugin depends on SetUserPasswordPlugin
and CreateUserPlugin in order to retrieve the credentials which will
be used for setting up the WinRM authentication. This is not desirable,
since ConfigWinRMCertificateAuthPlugin can't run independently
of the other two. This patch obtains the username from the config file,
if it is not found in the shared data dictionary and it retrieves
the password using get_admin_password from the metadata service, if it
is not found in the shared data dictionary.

Change-Id: I587782ff9cb8989bf01f6d86dbdbd1fa6ec53d1b
Closes-Bug: #1433174