Escape the password for WinRM certificate XML
This patch escapes the password formatted into a hand-crafted
XML for the WinRM certificate mapping. Without it, using passwords
such as 'P@ssw&d' will make the WinRM certificate auth plugin to
fail with a "The WS-Management service cannot process the request
because the XML is invalid.", which is actually true, since the resulting
XML is not valid, due to the unescaped and sign.

Change-Id: Ia93ab13a4ae5783c1fed5fbb748902bda84e9b65
Closes-Bug: #1441884