Add a mapping between old plugin names and new plugin names
Some plugins were moved to plugins.common, in order to
better reflect the fact that they are not platform specific.
Unfortunately, there are a lot of users out there with old
config files which are using the old plugin names.
So in order not to crash cloudbaseinit for their cases,
we provide an explicit mapping between old plugin names,
found in plugin.windows to new plugin names, found in plugins.common.
This will be removed when we'll reach 1.0 though.

Change-Id: I1102260849bbc940f2a26325c652e2cf02892eab