Check the latest version of cloudbaseinit
We're hitting an API with the current version and we expect to receive a
response if the current version is not the latest. This helps the user
to see that a new version of cloudbaseinit is available, which is a big
improvement than going to the github page to see the latest version.

Change-Id: Ibfb721973c84474c8fef8e1989dfb7566938134f