Fix mock_open issue in opennebula test
The state of the "open" mock object is not restored
when new data is read through the ".read" method,
even with the fake file object recreated. The only
possiblities are to make new open mock objects for
every read or to set a constant return value for
that particular mocked open.

Change-Id: Ibfa0edac24b9451b5bf9ba2500205353111758b3