Adds Windows Storage Manager API support
This is needed on Nano Server where VDS is not available.
The logic for extending the volumes was decoupled from the plugin to
a windows storage utility (VDS) alongside other extending utility which
is available on Nano and versions greater than ws2012/w8 (WSM).
According to the current Windows version, a suitable utility for extending
will be loaded through a factory manager.

Implements: blueprint windows-storage-manager
Co-Authored-By: Cosmin Poieana <cpoieana@cloudbasesolutions.com>
Change-Id: I7c8eea35a632c812e99808befb5e7528d66363cc