Add a new method `get_decoded_user_data` to the metadata services
The user data can be encoded with gzip, which means that every
access to it should verify this fact. Currently, the userdata
plugin is the sole consumer of it and it verified accordingly
if it was encoded or not, but there are situations (offline domain join),
where we need to inspect the user data from a different plugin. In order
to avoid code duplication, this method can be used to retrieve the decoded
user data. The original method wasn't modified to deal with this, because
there might be situations where we need the original user data unmodified
(and separations of concerns as a second reason).

Change-Id: I446ed293869af0749824ba0d386e1a58b0fe7c88