Updates tox.ini's cover environment
The openstack-tox-cover job expects a coverage report at the end.
This patch updates the cover environment in tox.ini, bringing it
in line with the other repos.

This also adds stestr in test-requirements, which is needed by the
new coverage env.

Finally, removes the cloudbaseinit import from releasenotes.
According to [1], a project shouldn't necessarely be installed
in order to build its releasenotes.

[1] http://lists.openstack.org/pipermail/openstack-dev/2017-November/124480.html

Change-Id: I77dfe57f1190fc62f7bd19284e2c86d55bb1804a