docs: add automatically generated list of cfg opts
This change will include an automatically generated list of
config options in the cloudbase-init docs.

It's useful as some config options are not currently mentioned in
the docs and it's also nice to be able to have a quick overview
of all available options without having to go through the code.
You can also send links to certain config options or groups.

Note that long config opt lines are not wrapped and look quite bad,
especially with the current theme (may consider switching to some
other theme, maybe the "alabaster" one included by default by
sphinx). Still better than not having this at all though.

We're also adding the zuul jobs that validate release notes and
docs, required for testing this patch.

Change-Id: Ibaff3a91e7f1345aaaac0c29b995e7c25b9e4669