Configure github mirroring
We'll mirror https://opendev.org/x/cloudbase-init to
https://github.com/cloudbase/cloudbase-init. The old
openstack/cloudbase repo has been transfered to the cloudbase
github organization.

In order to do this, we're adding a zuul Ansible job that will take
care of pushing the commits to the cloudbase mirror. The secret
defined there is actually a github deploy key encrypted with
zuul's public key.

Change-Id: I9b4869e0e7360750e47f2e7592a584c47d957129