Run cloud config plugins in order
The cloud config plugins were configured as a normal Python dictionary,
thus executing them in a random order, defined by the Python
implementation.

We require runcmd plugin to run last so that it is the last
configuration stage. In the future, we need the order as the groups
plugin is required to run before the users plugin.

Change-Id: I754ef7b15d5c7a12d08a12c876cfb52de64a06cc