Fix Py27 cloud-config ordering
On Py27, the yaml module implementation uses a normal dictionary for
implementation, resulting in arbitrary order for the parsed yaml.

If the config option cloud_config_plugins is not set or empty, fallback
to the default order defined in the factory.

Config option `cloud_config_plugins` is now used for filtering the cloud
config plugins to be executed, which was the original intended behaviour.

Change-Id: I59ea56e473ab343a826ff8c38787cb7f11cebd27