Fix Python 3 unit tests
Removed the functools partial from being mocked, as it is used
internally in the mocking implementation and generates unit test error.

Updated tox / unit test requirements according to new OpenStack gate
requirements.

Updating the tox requirements added a new set of flakes rules that had
to be skipped for now, which will be addressed in upcoming patches.

Change-Id: I231ca3b15313541e2c43785e80d9db5659aa1cc3