xml: force return of unicode string
With Python 3.8, the xml.etree.ElementTree.ElementTree function now
preserves the attribute order specified by the user.

This change broke the winrm enable functionality, throwing the erorr:

```
TypeError: Objects of type 'bytes' can not be converted to Unicode.
```

The fix was to set the correct parameter to return an unicode string and
not bytes.

Change-Id: I10f7fa357267f36dfed03ccb708694f6fc1cdd97