replace unittest.mock._get_target with mock one
Use the original importer for mock. Otherwise, the unit tests that
use importlib with context are not running in an isolated manner,
leading to various transient failures if `stestr` is not run with the
`--isolated` flag.

The unittest.mock._get_target has been changed in Python 3.11, by commit:
https://github.com/python/cpython/commit/ab7fcc8fbdc11091370deeb000a787fb02f9b13d

This change is needed to add support for Python 3.11 and Python 3.12
unit testing.

Change-Id: Ic73f5280116673f771f5460cdfd3702f5fbe16b5