unit_tests: fix winrm listener asserts
mock.Mock().method_name.return_value if not set specifically, it returns
a mock, thus being equivalent to a logical True.

Added a return value in either case, so that it is visible.
Without the return value set in both cases, the unit test fail on Python
3.12.

Also, the try-`finally` is run outside of the `with` context, thus
moving the `assert_has_calls` after `with` checks.

Change-Id: Ic33afb6d1403b1096e06406ffd6d36a969f823d3
Signed-off-by: Adrian Vladu <avladu@cloudbasesolutions.com>