Drop netifaces dependency
The netifaces library is no longer maintained, which is why we
need to drop this dependency.

There is only one place in which netifaces is being used, a
trivial function that retrives the mac address for a given ip
address.

Thankfully, we already have a cloudbase-init "get_adapter_addresses"
function that uses ctypes to call GetAdaptersAddresses, which
happens to be the same Windows function used by netifaces.

Worth mentioning that netifaces is the only compilable cloudbase-init
dependency that does not provide a wheel package.

Fixes: https://github.com/cloudbase/cloudbase-init/issues/140
Change-Id: Ie52ff722cbf42da7b9bfa9f9942adc1996ce5dd8